feat(nvim): telescope theme picker changes correctly config file
Before the file was not changed because, using any action_state would
change the in memory theme of the config value. That theme was then
used to compare with what was under the cursor and since they were no
differencce the replace_word, did not replace anything since the value
in memory can not be found in the config file.